test: add test

+++ lib/features/journal/ui/widgets/list_cards/modern_journal_image_card.dart
@@ -194,7 +194,7 @@ class ModernJournalImageCard extends StatelessWidget {
 
     return TextViewerWidgetNonScrollable(
       entryText: item.entryText,
-      maxHeight: availableHeight.toDouble(),
+      maxHeight: availableHeight,
     );
   }
